+++ udata/core/spatial/tests/test_fields.py
@@ -1,10 +1,11 @@
 import json
 from datetime import timedelta
 
+from mongoengine.fields import EmbeddedDocumentField
 from werkzeug.datastructures import MultiDict
 
 from udata.forms import Form
-from udata.mongo import db
+from udata.mongo.document import UDataDocument as Document
 from udata.tests import TestCase
 from udata.tests.helpers import assert_json_equal
 from udata.utils import faker
@@ -18,8 +19,8 @@ A_YEAR = timedelta(days=365)
 
 class SpatialCoverageFieldTest(TestCase):
     def factory(self):
-        class Fake(db.Document):
-            spatial = db.EmbeddedDocumentField(SpatialCoverage)
+        class Fake(Document):
+            spatial = EmbeddedDocumentField(SpatialCoverage)
 
         class FakeForm(Form):
             spatial = SpatialCoverageField()