+++ udata/settings.py
@@ -27,9 +27,9 @@ class Defaults(object):
     MONGODB_HOST = "mongodb://localhost:27017/udata"
     MONGODB_CONNECT = False  # Lazy connexion for Fork-safe usage
 
-    # Search service configuration
-    SEARCH_SERVICE_API_URL = None
-    SEARCH_SERVICE_REQUEST_TIMEOUT = 20
+    # Search configuration
+    ELASTICSEARCH_URL = None
+    ELASTICSEARCH_INDEX_BASENAME = None
 
     # BROKER_TRANSPORT = 'redis'
     CELERY_BROKER_URL = "redis://localhost:6379"
@@ -191,8 +191,6 @@ class Defaults(object):
     TERMS_OF_USE_URL = None
     TERMS_OF_USE_DELETION_ARTICLE = None
     TELERECOURS_URL = None
-
-    UDATA_INSTANCE_NAME = "udata"
 
     DATASET_HIDDEN_BADGES = []
 
@@ -528,6 +526,8 @@ class Defaults(object):
     URLS_ALLOW_PRIVATE = False
     # Whether or not to allow local URLs (localhost...) submission.
     URLS_ALLOW_LOCAL = False
+    # Wheter or not to resolve hostname to verify for private or local IP.
+    URLS_RESOLVE_HOSTNAME = True
     # Whether or not to allow credentials in URLs submission.
     URLS_ALLOW_CREDENTIALS = True
     # List of allowed URL schemes.
@@ -698,6 +698,7 @@ class Testing(object):
     LOGGER_HANDLER_POLICY = "never"
     CELERYD_HIJACK_ROOT_LOGGER = False
     URLS_ALLOW_LOCAL = True  # Test server URL is local.test
+    URLS_RESOLVE_HOSTNAME = False
     URLS_ALLOWED_TLDS = tld_set | set(["test"])
     URLS_ALLOW_PRIVATE = False
     FS_IMAGES_OPTIMIZE = True
@@ -707,9 +708,13 @@ class Testing(object):
     SECURITY_TWO_FACTOR = True  # should be set before security init_app for views to be loaded
     PUBLISH_ON_RESOURCE_EVENTS = False
     HARVEST_ACTIVITY_USER_ID = None
-    SEARCH_SERVICE_API_URL = None
     CDATA_BASE_URL = None
     SCHEMA_CATALOG_URL = None
+    SPAM_WORDS = []
+    SPAM_ALLOWED_LANGS = []
+    DATASET_HIDDEN_BADGES = []
+    ELASTICSEARCH_URL = None
+    ELASTICSEARCH_INDEX_BASENAME = "udata-test"
 
 
 class Debug(Defaults):