+++ udata/tests/forms/test_current_user_field.py
@@ -1,6 +1,7 @@
-import datetime
+from datetime import UTC, datetime
 
 from bson import ObjectId
+from mongoengine.fields import ReferenceField
 from werkzeug.datastructures import MultiDict
 
 from udata.auth import login_user
@@ -8,15 +9,16 @@ from udata.auth.forms import ExtendedLoginForm, ExtendedRegisterForm
 from udata.core.user.factories import AdminFactory, UserFactory
 from udata.forms import ModelForm, fields
 from udata.i18n import gettext as _
-from udata.models import User, db
+from udata.models import User
+from udata.mongo.document import UDataDocument as Document
 from udata.tests.api import DBTestCase
 from udata.tests.helpers import security_gettext
 
 
 class CurrentUserFieldTest(DBTestCase):
     def factory(self, *args, **kwargs):
-        class Ownable(db.Document):
-            owner = db.ReferenceField(User)
+        class Ownable(Document):
+            owner = ReferenceField(User)
 
         class OwnableForm(ModelForm):
             model_class = Ownable
@@ -180,7 +182,7 @@ class CurrentUserFieldTest(DBTestCase):
         self.assertEqual(len(form.errors["owner"]), 1)
 
     def test_password_rotation(self):
-        today = datetime.datetime.utcnow()
+        today = datetime.now(UTC)
         user = UserFactory(
             password="password", password_rotation_demanded=today, confirmed_at=today
         )
@@ -214,7 +216,7 @@ class CurrentUserFieldTest(DBTestCase):
         form.validate()
         self.assertIn(security_gettext("Invalid email address"), form.errors["email"])
 
-        today = datetime.datetime.utcnow()
+        today = datetime.now(UTC)
         user = UserFactory(email="b@fake.com", password="password", confirmed_at=today)
         form = ExtendedLoginForm.from_json({"email": user.email, "password": "password"})
         form.validate()