+++ udata/tests/forms/test_image_field.py
@@ -2,11 +2,12 @@ import logging
 
 import flask_storage as fs
 import pytest
+from flask_storage.mongo import ImageField as MongoImageField
 
 from udata.core.storages import tmp
 from udata.forms import Form
 from udata.forms.fields import ImageField
-from udata.mongo import db
+from udata.mongo.document import UDataDocument as Document
 from udata.tests.api import PytestOnlyDBTestCase
 from udata.tests.helpers import data_path
 
@@ -26,9 +27,9 @@ class PostData(dict):
 
 
 class ImageFieldTest(PytestOnlyDBTestCase):
-    class D(db.Document):
-        image = db.ImageField(fs=storage)
-        thumbnail = db.ImageField(fs=storage, thumbnails=SIZES)
+    class D(Document):
+        image = MongoImageField(fs=storage)
+        thumbnail = MongoImageField(fs=storage, thumbnails=SIZES)
 
     class F(Form):
         image = ImageField()