+++ udata/tests/test_discussions.py
@@ -1,4 +1,4 @@
-from datetime import datetime
+from datetime import UTC, datetime
 
 import pytest
 from flask import url_for
@@ -372,7 +372,7 @@ class DiscussionsTest(APITestCase):
                 user=user,
                 title="test discussion {}".format(i),
                 discussion=[message],
-                closed=datetime.utcnow(),
+                closed=datetime.now(UTC),
                 closed_by=user,
             )
             closed_discussions.append(discussion)
@@ -404,7 +404,7 @@ class DiscussionsTest(APITestCase):
                 user=user,
                 title="test discussion {}".format(i),
                 discussion=[message],
-                closed=datetime.utcnow(),
+                closed=datetime.now(UTC),
                 closed_by=user,
             )
             closed_discussions.append(discussion)
@@ -1140,7 +1140,7 @@ class NotifyDiscussionsTest(APITestCase):
             user=poster,
             title=faker.sentence(),
             discussion=[message, second_message, closing_message],
-            closed=datetime.utcnow(),
+            closed=datetime.now(UTC),
             closed_by=owner,
         )